🚚 Move web arbitraries into /arbitrary
